Extract Access Tokens into their own controller
