+++ app/components/my/access_token/new_access_token_form_component.html.erb
@@ -31,7 +31,7 @@ See COPYRIGHT and LICENSE files for more details.
   component_wrapper do
     primer_form_with(
       model: @token,
-      url: my_generate_api_key_path,
+      url: generate_api_key_my_access_tokens_path,
       data: { turbo: true },
       method: :post
     ) do |form|
