+++ app/views/my/access_tokens/_icalendar_meeting_tokens_section.html.erb
@@ -26,22 +26,22 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
-<div class="attributes-group" id="rss-token-section">
+<div class="attributes-group" id="icalendar-token-section">
   <div class="attributes-group--header">
     <div class="attributes-group--header-container">
-      <h3 class="attributes-group--header-text"><%= t("my_account.access_tokens.rss.title") %></h3>
+      <h3 class="attributes-group--header-text"><%= t("my_account.access_tokens.ical_meeting.title") %></h3>
     </div>
   </div>
   <div>
     <p>
-      <%= t("my_account.access_tokens.rss.text_hint") %>
+      <%= t("my_account.access_tokens.ical_meeting.text_hint") %>
     </p>
-    <% if Setting.feeds_enabled? %>
-      <% if rss_token %>
+    <% if Setting.ical_enabled? %>
+      <% if ical_tokens.any? %>
         <div class="generic-table--container">
           <div class="generic-table--results-container">
             <table class="generic-table" data-controller="table-highlighting">
-              <%= render  partial: "my/access_token_partials/token_table_header",
+              <%= render  partial: "token_table_header",
                           locals: {
                             column_headers: [
                               t("attributes.name"),
@@ -50,37 +50,37 @@ See COPYRIGHT and LICENSE files for more details.
                             ]
                           } %>
               <tbody>
-                <tr>
-                  <td class="-w-rel-60" data-test-selector="rss-token-row-name"><%= t("my_account.access_tokens.rss.static_token_name") %></td>
-                  <td>
-                    <span title="<%= format_time(rss_token.created_at) %>">
-                      <%= format_time(rss_token.created_at.to_s) %>
-                    </span>
-                  </td>
-                  <td><%= t("my_account.access_tokens.indefinite_expiration") %></td>
-                  <td class="buttons">
-                    <%= link_to "",
-                                { action: "revoke_rss_key" },
-                                method: :delete,
-                                data: { confirm: t("my_account.access_tokens.simple_revoke_confirmation"), test_selector: "rss-token-revoke" },
-                                class: "icon icon-delete" %>
-                  </td>
-                </tr>
+                  <% ical_tokens.each do |token| %>
+                    <tr>
+                      <td class="-w-rel-20 -mw-abs-200" data-test-selector="ical-meeting-token-row-<%= token.id %>-name"><%= token.token_name %></td>
+                      <td>
+                        <span title="<%= format_time(token.created_at) %>">
+                          <%= format_time(token.created_at.to_s) %>
+                        </span>
+                      </td>
+                      <td><%= t("my_account.access_tokens.indefinite_expiration") %></td>
+                      <td class="buttons">
+                        <%= link_to "",
+                                    { action: "revoke_ical_meeting_token", id: token.id },
+                                    data: { confirm: t("my_account.access_tokens.simple_revoke_confirmation"), test_selector: "ical-meeting-token-row-#{token.id}-revoke" },
+                                    method: :delete,
+                                    class: "icon icon-delete" %>
+                      </td>
+                    </tr>
+                <% end %>
               </tbody>
             </table>
           </div>
         </div>
       <% else %>
-        <%= link_to({ action: "generate_rss_key" }, method: :post, class: "spot-link", data: { test_selector: "rss-token-add" }) do %>
-          <span class="spot-icon spot-icon_add"></span>
-          <span><%= t("my_account.access_tokens.rss.static_token_name") %></span>
-        <% end %>
+        <span class="spot-icon spot-icon_inline spot-icon_info1"></span>
+        <i><%= t("my_account.access_tokens.ical_meeting.empty_text_hint", path: meetings_path).html_safe %></i>
       <% end %>
     <% else %>
       <div tabindex="0" class="-info op-toast">
         <div role="alert" aria-atomic="true" class="op-toast--content">
           <p>
-            <span><%= t("my_account.access_tokens.rss.disabled_text") %></span>
+            <span><%= t("my_account.access_tokens.ical_meeting.disabled_text") %></span>
           </p>
         </div>
       </div>
