+++ app/views/my/access_tokens/_oauth_tokens_section.html.erb
@@ -40,7 +40,7 @@ See COPYRIGHT and LICENSE files for more details.
       <div class="generic-table--container">
         <div class="generic-table--results-container">
           <table class="generic-table" data-controller="table-highlighting">
-            <%= render  partial: "my/access_token_partials/token_table_header",
+            <%= render  partial: "token_table_header",
                         locals: {
                           column_headers: [
                             t("attributes.name"),
